+++ lib/bigml/fields.rb
@@ -28,7 +28,8 @@ module BigML
   RESOURCES_WITH_FIELDS = [SOURCE_PATH, DATASET_PATH, MODEL_PATH,
                            PREDICTION_PATH, CLUSTER_PATH, ANOMALY_PATH,
                            SAMPLE_PATH, CORRELATION_PATH, STATISTICAL_TEST_PATH,
-                           LOGISTIC_REGRESSION_PATH, ASSOCIATION_PATH, TOPIC_MODEL_PATH]
+                           LOGISTIC_REGRESSION_PATH, ASSOCIATION_PATH, 
+                           TOPIC_MODEL_PATH, ENSEMBLE_PATH]
 
   DEFAULT_MISSING_TOKENS = ["", "N/A", "n/a", "NULL", "null", "-", "#DIV/0",
                             "#REF!", "#NAME?", "NIL", "nil", "NA", "na",
@@ -82,6 +83,8 @@ module BigML
           fields = resource['associations']['fields']
       elsif resource_type == TOPIC_MODEL_PATH
           fields = resource['topic_model']['fields']
+      elsif resource_type == ENSEMBLE_PATH
+          fields = resource['ensemble']['fields']
       elsif resource_type == SAMPLE_PATH
           fields = {}
           resource['sample']['fields'].each do |field|